refactor hotkey methods to central utility
