+++ .gitignore
@@ -49,3 +49,4 @@ build
 .env
 **/ai-decision-debugging/data/*
 apps/astriarch-backend/analysis/*
+.startbuilding/runs/
